Move toolbar to apputils
